Giving the logger names same scheme
