+++ lib/java/jvm-interop/src/main/java/org/enso/jvm/interop/impl/OtherJvmUtils.java
@@ -76,7 +76,7 @@ public final class OtherJvmUtils {
     commandAndArgs.add("-Djdk.module.main=" + mainModule);
     // commandAndArgs.add("-Djdk.module.showModuleResolution=true");
     commandAndArgs.add(
-        "-Dslf4j.provider=org.enso.logging.config.systemlogger.SystemLoggerForSlf4j");
+        "-Dslf4j.provider=org.enso.logging.config.systemlogger.Slf4ViaSystemProvider");
     commandAndArgs.add("-Dslf4j.internal.verbosity=WARN");
     return commandAndArgs.toArray(String[]::new);
   }
