+++ lib/scala/logging-config/src/main/java/org/enso/logging/config/systemlogger/Slf4ViaSystemProvider.java
@@ -10,7 +10,7 @@ import org.slf4j.spi.SLF4JServiceProvider;
  * Bridges slf4j logger to {@link System.Logger}. Needs to be enabled by setting JVM properties.
  * Those are set when booting <em>Other JVM</em> in the <em>dual JVM</em> mode.
  */
-public final class SystemLoggerForSlf4j implements SLF4JServiceProvider, ILoggerFactory {
+public final class Slf4ViaSystemProvider implements SLF4JServiceProvider, ILoggerFactory {
 
   @Override
   public ILoggerFactory getLoggerFactory() {
@@ -38,6 +38,6 @@ public final class SystemLoggerForSlf4j implements SLF4JServiceProvider, ILogger
   @Override
   public Logger getLogger(String name) {
     var delegate = System.getLogger(name);
-    return new SystemSlf4jLogger(delegate);
+    return new Slf4jViaSystemLogger(delegate);
   }
 }
